// projects/mt-select-dropdown/src/lib/pipes/pipes/types.ts
export interface StyleClassOptions {
    dropdown_container?: string;

    // Styling options for the custom input div
    dropdown_form_control?: string;

    // Styling options for the selected options container in multi-select mode
    selectedContainer?: string;

    // Styling options for the selected option item in multi-select mode
    selectedItem?: string;

    // Styling options for the unselect item (cross) in multi-select mode
    unselectItem?: string;

    // Styling options for the placeholder text
    placeholder?: string;

    // Styling options for the drop symbol
    dropSymbol?: string;

    // Styling options for the dropdown menu
    dropdownMenu?: string;

    // Styling options for the search input in the dropdown
    searchInput?: string;

    // Styling options for the dropdown items in the list
    dropdownItem?: string;
}